requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
bids-validator==1.9.9
brainspace==0.1.4
certifi==2022.9.24
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
ConfigArgParse==1.5.3
contourpy==1.0.5
cycler==0.11.0
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
docopt==0.6.2
etelemetry==0.3.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.8.0
Flask==2.2.3
fonttools==4.38.0
formulaic==0.3.4
frozenlist==1.3.3
h5py==3.7.0
idna==3.4
imageio==2.22.2
importlib-metadata==6.5.0
importlib-resources==5.12.0
interface-meta==1.3.0
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.0.3
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
looseversion==1.0.2
lxml==4.9.1
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mriqc-learn==0.0.2
multidict==6.0.2
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
nibabel==3.2.2
nilearn==0.9.2
nipype==1.8.5
nitransforms==21.0.1
niworkflows==1.5.4
num2words==0.5.12
numpy==1.23.4
open3d==0.17.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
plotly==5.14.1
prompt-toolkit==3.0.38
prov==2.0.0
psutil==5.9.5
pure-eval==0.2.2
pybids==0.15.5
pydicom==2.3.0
pydot==1.4.2
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
rdflib==6.2.0
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
SimpleITK==2.2.0
simplejson==3.18.0
six==1.16.0
SQLAlchemy==1.3.24
stack-data==0.6.2
statsmodels==0.13.5
surfplot==0.1.0
svgutils==0.3.4
templateflow==0.8.1
tenacity==8.2.2
threadpoolctl==3.1.0
tifffile==2022.10.10
toml==0.10.2
torch==1.8.1+cpu
torchvision==0.9.1+cpu
tornado==6.3
tqdm==4.64.1
traitlets==5.9.0
traits==6.3.2
transforms3d==0.4.1
typing_extensions==4.4.0
urllib3==1.26.12
vtk==9.2.2
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.14.1
wslink==1.9.1
yarl==1.8.1
zipp==3.15.0
